requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
firebase-admin==6.2.0
pykrx==1.0.45
pandas>=2.0.0
openai>=1.0.0
bcrypt>=4.0.0
